% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psx.R
\name{psx}
\alias{psx}
\title{shorthand wrapper for as.POSIXct}
\usage{
psx(x, tz = "UTC", ...)
}
\arguments{
\item{x}{input}

\item{tz}{timezone}

\item{...}{additional arguments}
}
\description{
shorthand wrapper for as.POSIXct
}
